requirements.txt
PyQt5~=5.15.7
fonttools~=4.38.0
Pillow~=9.3.0
